% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirads.R
\name{TIRADS_calc}
\alias{TIRADS_calc}
\title{ACR TIRADS Calculator}
\usage{
TIRADS_calc(
  composition,
  echogenicity,
  shape,
  margin,
  echogenic_foci,
  size_cm = NA
)
}
\arguments{
\item{composition}{(character) Internal components of a nodule,
must be one of "cystic", "spongiform","mixed", or "solid".}

\item{echogenicity}{(character) Level of echogenicity of the solid part,
relative to surrounding thyroid tissue. One of: "an" (anechoic), "hyper" (hyperechoic),
"iso" (isoechoic), "hypo" (hypoechoic), or "very-hypo" (very hypoechoic).}

\item{shape}{(character) Shape of the nodule, one of: "wider" (wider-than-tall) or "taller" (taller-than-wide).}

\item{margin}{(character) Margin of the nodule, one of: "smooth", "ill-defined", "lobulated", "irregular",
or "extra" (extrathyroidal extension)}

\item{echogenic_foci}{(character) Can choose > 1: "none", "comet" (large comet-tail artifacts), "macro-calc" (macrocalcification),
"rim-calc" (rim calcification), or "punctate" (punctate echogenic foci).}

\item{size_cm}{(Numeric) Size of the nodule in cm.}
}
\value{
A list with "TIRADS_report" class containing:
\itemize{
\item \strong{level}: (Character) A TIRADS category level.
\item \strong{tot_points}: (Numeric) A total TIRADS score.
\item \strong{should_FNA}: (Logical) \code{TRUE} if FNA is suggested.
\item \strong{should_follow}: (Logical) \code{TRUE} if follow-up is suggested.
}
}
\description{
Calculate \href{https://www.acr.org/Clinical-Resources/Reporting-and-Data-Systems/TI-RADS}{ACR TIRADS} score, category level, and whether follow up or FNA should be done.
}
\seealso{
\itemize{
\item \href{https://www.acr.org/Clinical-Resources/Reporting-and-Data-Systems/TI-RADS}{ACR TIRADS (Thyroid Imaging Reporting & Data System)}
}
}
